fix(temporal): batch-3 review fixes — type alignment, SAFETY docs, load bounds
- storage_types: change HotspotRow.changes_30d/changes_90d, RiskRow.total_commits/
  fix_commits, CochangeRow.count from i64 to u32, aligning with FileTemporalStats
  in types.rs; rusqlite 0.31 implements ToSql/FromSql for u32 so insert/load
  helpers need no explicit widening
- storage_ops: correct SAFETY comments on store_hotspots and sync — rusqlite
  Connection IS Send (just not Sync); the invariant is 'Send but not Sync'
- storage_ops: add LIMIT 500001 and post-load CapacityExceeded guard to
  load_hotspots, load_risks, load_cochanges, matching store/sync enforcement
- storage_perf_tests: update make_hotspot/make_risk/make_cochange helpers to
  use plain u32 casts instead of i64::try_from

+++ crates/rskim-search/src/temporal/storage_types.rs
@@ -15,9 +15,9 @@ pub struct HotspotRow {
     /// Decay-weighted commit frequency, max-normalized to `[0.0, 1.0]`.
     pub score: f64,
     /// Raw commit count within the last 30 days.
-    pub changes_30d: i64,
+    pub changes_30d: u32,
     /// Raw commit count within the last 90 days.
-    pub changes_90d: i64,
+    pub changes_90d: u32,
 }
 
 /// A row from the `risk` table.
@@ -28,9 +28,9 @@ pub struct RiskRow {
     /// Bug-fix density score in `[0.0, 1.0]`.
     pub risk_score: f64,
     /// Total number of commits touching this file.
-    pub total_commits: i64,
+    pub total_commits: u32,
     /// Number of commits classified as fix commits.
-    pub fix_commits: i64,
+    pub fix_commits: u32,
     /// Ratio of fix commits to total commits, in `[0.0, 1.0]`.
     pub fix_density: f64,
 }
@@ -43,7 +43,7 @@ pub struct CochangeRow {
     /// Repository-root-relative path of the second file in the pair.
     pub file_b: String,
     /// Number of commits that touched both files.
-    pub count: i64,
+    pub count: u32,
     /// Jaccard similarity of the two files' commit sets.
     pub jaccard: f64,
 }
